Refactor and cleanup code
This change moves some code around to better separate areas of concern
and removes variables that are not necessary.

One note: the code that is moved here is not changed in any way. It is
only moved from one file to another.
